some fixes

+++ ShopTest/Data/Entities/Product.cs
@@ -9,10 +9,10 @@ namespace ShopTest.Data.Entities
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public long Id { get; set; }
         public string Name { get; set; }
-        public Decimal Price { get; set; }
+        public Double Price { get; set; }
         public int Count { get; set; }
         [DatabaseGenerated(DatabaseGeneratedOption.Computed)]
-        public Decimal TotalPrice { get; private set; }       
+        public Double TotalPrice { get; private set; }       
         public Category Category { get; set; }
         public ICollection<Purchase> Purchase { get; set; }
     }
